requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.6.0
    # via django
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
django==4.2.1
    # via -r requirements.in
idna==3.4
    # via requests
marshmallow==3.17.0
    # via transbank-sdk
mysqlclient==2.1.1
    # via -r requirements.in
packaging==23.1
    # via
    #   marshmallow
    #   plotly
plotly==5.14.1
    # via -r requirements.in
requests==2.31.0
    # via transbank-sdk
sqlparse==0.4.4
    # via django
tenacity==8.2.2
    # via plotly
transbank-sdk==4.0.0
    # via -r requirements.in
tzdata==2023.3
    # via django
urllib3==2.0.2
    # via requests
